stock_analyzer_LangGraph/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
openai
yfinance
pandas
pydantic
google-serper-api
python-dotenv
typing
json
IPython